netdev_driver: get/set power with signed values

+++ stack/framework/hal/chips/netdev_driver/sx127x_netdev.c
@@ -475,8 +475,8 @@ static int _get(netdev_t *netdev, netopt_t opt, void *val, size_t max_len)
             return sizeof(netopt_enable_t);
 
         case NETOPT_TX_POWER:
-            assert(max_len >= sizeof(int16_t));
-            *((int16_t*) val) = (int16_t)sx127x_get_tx_power(dev);
+            assert(max_len >= sizeof(int8_t));
+            *((int8_t*) val) = sx127x_get_tx_power(dev);
             return sizeof(int16_t);
 
         case NETOPT_IQ_INVERT:
@@ -671,8 +671,8 @@ static int _set(netdev_t *netdev, netopt_t opt, const void *val, size_t len)
                 res = -EINVAL;
                 break;
             }
-            sx127x_set_tx_power(dev, (int8_t)power);
-            return sizeof(int16_t);
+            sx127x_set_tx_power(dev, power);
+            return sizeof(int8_t);
 
         case NETOPT_FIXED_HEADER:
             assert(len <= sizeof(netopt_enable_t));
